Cleanup of duplicate api files.
